refactor: gameTime 패키지 수정, 메서드명 수정
